+++ janus-admin-app/src/main/java/org/xujin/janus/app/command/query/QueryClusterCountCmdExe.java
@@ -1,24 +1,13 @@
 package org.xujin.janus.app.command.query;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.google.common.collect.Lists;
-
 import org.xujin.halo.annotation.command.CmdHandler;
 import org.xujin.halo.command.CommandExecutorI;
 import org.xujin.halo.dto.ResultData;
-import org.xujin.halo.exception.BusinessException;
 import org.xujin.janus.app.command.cmo.query.QueryClusterCountCmd;
-import org.xujin.janus.domain.user.service.ClusterService;
-import org.xujin.janus.domain.user.service.UserService;
-import org.xujin.janus.infrastructure.ClusterDetailCo;
 import org.xujin.janus.infrastructure.common.HaloConstant;
 import org.xujin.janus.infrastructure.tunnel.db.dao.ClusterMapper;
 import org.xujin.janus.infrastructure.tunnel.db.dataobject.ClusterDO;
-import org.xujin.janus.infrastructure.utils.SessionUtils;
-import io.swagger.models.auth.In;
-import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.annotation.Resource;
 
@@ -29,19 +18,19 @@ import javax.annotation.Resource;
  * @date 2020/6/11 14:34
  **/
 @CmdHandler
-public class QueryClusterCountCmdExe implements CommandExecutorI<ResultData<Integer>, QueryClusterCountCmd> {
+public class QueryClusterCountCmdExe implements CommandExecutorI<ResultData<Long>, QueryClusterCountCmd> {
 
     @Resource
     private ClusterMapper clusterMapper;
 
     @Override
-    public ResultData<Integer> execute(QueryClusterCountCmd cmd) {
-        Integer result = 0;
+    public ResultData<Long> execute(QueryClusterCountCmd cmd) {
+        Long result = 0L;
         result = countCluster();
         return ResultData.success(result);
     }
 
-    public Integer countCluster() {
+    public Long countCluster() {
         QueryWrapper<ClusterDO> queryWrapper = new QueryWrapper<>();
         queryWrapper.eq("is_deleted", HaloConstant.IS_DELETED_FALSE);
         return clusterMapper.selectCount(queryWrapper);
